Flip DParam and DParamNRel, so NRel is the default
Summary: Only LdLoc and LdStack need this.  The LdLocPseudoMain support didn't
seem to be entirely working: when we gen'd new LdLocPseudoMains, we'd map their
type sources to whatever the current type source was, and LdLocPseudoMain would
set its own dst as a type source.  Right now it's also setting types in
localTypes which are incorrect without a new guard (this could cause problems
for any callers of localType, but I don't know if it can just be removed, so I
made a task).  So any constrainValue on a LdLocPseudoMain would constrain the
/previous/ LdLocPseudoMain in the tracelet.  However, isGuardOp returns false
for these, so it can't have worked.  Also I don't understand why
StLocPseudoMain was modifying type srcs, so just remove that too.

Reviewed By: @markw65

Differential Revision: D1687087

Signature: t1:1687087:1416589588:d25561f895d2c6b994366485dc372493e0d6ff9e