Avoid saving global storage section.
Summary: I took a closer look at the code and realized that the 1GB of global
storage is always reset after each typechecking phase. (The memory isn't
zeroed out, but it is marked as unused.)

So this commit moves the global storage section to the front of the shared
memory (instead of in the middle), and hh_server only saves everything that
comes after that.

Reviewed By: @elgenie

Differential Revision: D1683934

Signature: t1:1683934:1416280066:0567f4d4ec45d72b4670851a51f413b693fdf3ce