Simplify coverage JSON output, and add non-JSON output option.
Summary: It is easy enough for the scripts that consume the JSON to do
the calculation of the "score" themselves. There are several ways to
calculate the score, and having hh calculate one of them is
unnecessarily restrictive.

So now the JSON output is of the form

    { "type": "file" | "directory",
      "result": { "checked": int, "unchecked": int, "partial": int },
      "children": <only present if type is directory>
    }

However, for the command-line output, it is still worthwhile to
calculate some scoring metrics since this output may consumed directly
by a user and not a script.

I ran into some trouble with getting the OSS build to compile initially;
ocamlc did not seem to understand that `ServerCoverage.CL` and
`ClientCoverage.CL` were one and the same module, so I replaced all the
module aliases by module opens. That required me to choose better names
for the declarations in the Coverage_level module so that opening the
module would not introduce confusing names like `t` into scope. I think
the code looks cleaner after the refactor too.

Reviewed By: @jwatzman

Differential Revision: D1684486

Signature: t1:1684486:1416274590:feaae6416ab9d939650b72ea75b71f792619a231