Update generate-buildinfo.sh for mercurial
Summary: This uses the manifest to generate a hash based on
the file contents + the output of "hg diff -- hphp".

This isn't quite right, because committing any changes will
change the hash, even though the file contents haven't changed,
but is better than using the time of day.

Also got rid of unwanted output from hhvm-land.

Reviewed By: @jasone

Differential Revision: D1687893