refactor smartMallocSizeBig
Summary: For jemalloc, it uses a never-inlined helper. otherwise, it
is layered on top of smartMallocBig, and I wanted to break
the dependency of the size-tracked allocator api on the guts
of smart_malloc. (ultimately, so big size-tracked don't use
BigNode at all).

1. make smartMallocBigHelper return a pair<void*,size_t> instead
of returning ptr & size by ref. (the ptr-by-ref arg was unused!)

2. move smartMallocSizeBig to memory-manager-inl.cpp, since both
paths would make a call anyway. (I'm counting on big allocations
not being very hot).

3. inline smartMallocSizeBigHelper into the jemalloc version of
smartMallocSizeBig

4. Factor out common code so just the different parts are #ifdef'd.
It would be nice to use if(use_jemalloc) instead of #ifdef, but
that would break the non-jemalloc build.

Reviewed By: @bertmaher

Differential Revision: D1690829

Signature: t1:1690829:1416421088:4b692cdabc235eb8b6a223a9de8cbb211679c749