Kill GlobalNameValueTableWrapper, do size-tracked allocation.
Summary: GNVTW is the only way we ever instantiate an NVTW, so flatten
them together. Also, allocate the one-instance-per-request
using the size-tracked api, since a) we know it's size, and
b) this exposes it to the heap parser so its not hidden in
behind a smart_malloc header.

I will rename NVTW to GlobalsArray in a followon diff.

Reviewed By: @jdelong

Differential Revision: D1691188