Some FrameState refactoring and bug fixes
Summary: Turn all the states into vectors, with the most current version being
the back().  This makes it easier to merge states for inlined calls and also
simplifies various snapshotting code (assignment operators).  Also it gets rid
of the weird thing where the Snapshot structure had a sometimes-valid
inlineSavedStates field.  This diff fixes the potential bugs from things like
refineLocalValue when you're in an inlined function with control flow, and
fixes a bug in the fixed point version of frame-state where frameSpansCall
could randomly be set to true depending on which block you were previously
visiting.

This is an intermediate refactor, so things like the name 'Snapshot' will go
away on a later diff.

Reviewed By: @bertmaher

Differential Revision: D1695209