Fix insertPhis calls to IRUnit::replace()
Summary: It was trying to replace a jump but passing no destination block.  Due
to a bug in IRInstruction::become this actually worked, because it left the old
Edges in place.  It also seems really sketchy to iterate the predecessor
EdgeList while you're modifying instructions that change it; I'm not sure if
that's ok (it probably could be if we guarantee not to move the Edge*'s if it
already had Edges, but that seems like a weird thing to rely on), so stop doing
it.  The function does something pretty similar a few lines earlier while
splitting critical edges also, but this part is explicitly doing setTaken and
setNext calls without anything that could be changing the Edge*, which seems
like it /might/ work, so I'm going to leave that part alone for now.

Reviewed By: @bertmaher

Differential Revision: D1693126

Signature: t1:1693126:1416451300:c4c3dafb5de2c6b92828fb4dd4ecc359e0fdb6cd