Add possibility to create a DateTime object with timezone
Summary: or the class will first initialize an object with a timezone, issue the strict warning and the timezone could be changed later
We don't want the warning.

The slow test didn't match php before but does now.
The zend test is clowny in my opinion. It tests for error position, that should have wildcards.
We mathc php7 in this case.

Closes #3650

Closes https://github.com/facebook/hhvm/pull/4212

Reviewed By: @paulbiss

Differential Revision: D1673795

Signature: t1:1673795:1416502867:dd4c4279f2d533b52685b9ee65557fb0628288df