Move CSE-related code out of FrameStateMgr
Summary: IRBuilder is not a good final place for it either, but this gets it
out of the way of some of the frame state stuff for now.  It's apparently
semi-broken during reoptimize (whether it finds values or not depends on the
block visitation order; e.g. you may cseKill values on some paths that you
don't really need to just because you visited blocks in an order that saw
killing-paths first).  I think LdCtx is the only important thing we CSE right
now though, and it'll always be available now.

Reviewed By: @alexmalyshev

Differential Revision: D1695435

Signature: t1:1695435:1416613090:df2fc292516bb651d67f4c5c2e990feb028829cd