Better fix for generate-buildinfo.sh
Summary: I had been assuming that computing a full hash of
all the involved files would be too slow. It turns out
its not too bad unless the cache is cold; but most of the
files are going to be read during the build anyway, so this
just preseeds the cache anyway.

This method has the advantage that it produces the same repo-schema
for git and hg if you have the same rev checkout out (and even with
no source control, as long as there aren't any "stray" files lying
around).

Reviewed By: @jasone

Differential Revision: D1694904