Don't track incorrect types for locals in psuedomains
Summary: Right now we have localType() returning things that aren't known to be
true in psuedomains.  Rather than audit all callers (which will still let bugs
creep in later), let's generalize the predicted inner type for boxes to a
general predicted type, and use it for the only real type we track for locals
in psuedomains.  Use this to decide what to guard on when we want to emit a
LdLocPseudoMain.

This also fixes a bug where we were doing StLoc in psuedomains in the Silence
opcode (needs to be illegal because it'll set a tracked local value right now).

Reviewed By: @alexmalyshev

Differential Revision: D1696464