Move LocalStateHook driving out of FrameStateMgr
Summary: This lets refcount opts talk to it directly.  But it still requires a
FrameStateMgr& for now to be able to call MInstrEffects::get (which wants to
read the old type of a local on the other side of a pointer, and is assuming
FrameStateMgr knows what that type is, although it only kinda does (not really
sound after gen time)).  Also gets rid of forEachFrame (unused) and makes the
forEachLocal callback match how it's used.

walkAllInlinedLocals also only has a single callsite now, which needs it
because of the skipFrame stuff.  (You can't null the locals out more than once
or refcount opts thinks it's consuming things too many times.)

Reviewed By: @swtaarrs

Differential Revision: D1699847