Fix frameSpansCall bugs, remove 'clearing' code and add m_status asserts
Summary: There was no merging at all of the frameSpansCall state, and the way
we handled it at a loop header or with an unprocessed pred was incorrect (it
left it dangling as whatever it used to be).  Rename it to make it clearer that
it's may-information, and refactor the various clear/reset state functions so
they don't exist anymore.  Expanded m_status with some assertions: it's pretty
clear with the status assertions that this class is still doing 3 unrelated
things (finding a fixed point, dag-only-legacy-reoptimize-or-analyze-stuff, and
the building mode), so it should be split up further.

Reviewed By: @swtaarrs

Differential Revision: D1699849

Signature: t1:1699849:1416856507:403366c23fa1471eaa1ded1486725ed93b83ad15