Remove runtime/test/refcount-opts.cpp
Summary: It's the only user of FrameStateMgr's second constructor, which I want
to remove, and you can't pass a dummy marker to the normal constructor instead
as we end up hitting hasFunc assertions (which looks like it's explicitly
testing that it's not a dummy).  I could make it work (or leave the
constructor), but I think this test's days are numbered anyway, since we're
going to need to change how the refcount-opts pass works pretty fundamentally
for it to support arbitrary control flow.  So unless someone really likes it
can I delete it?

Reviewed By: @swtaarrs

Differential Revision: D1699848

Signature: t1:1699848:1416854866:70d6f636671ac0dd6c46773702e296b2d7003bc9