Remove marker from FrameState
Summary: It seemed weird to duplicate spOff and curFunc with the marker, and
then I was a little worried that it isn't handle the same way for merging.  It
turns out it wasn't broken, just not necessary: ir-builder is just storing its
next-to-gen-marker there for presumably historical reasons.

This also changes splitEdge to use the marker from the source jump instead of
the destination, which seems to make more sense to me, but I think we could
choose either.

Reviewed By: @alexmalyshev

Differential Revision: D1699852

Signature: t1:1699852:1416889935:4735fb88a4ec8ed3f24a4d5937275e3fa7d2319b