Allow null to password_hash()
Summary: PHP 5.6 allows `null` to be passed to `password_hash()`. Now we can pass it.

This is only a partial fix for #4294 since `password_hash()` also allows `int`s that can be
converted to `string` as well. Instead of making the parameter a `mixed`, we will wait for
`ParamCoerceMode` (PCM) to land, which should take care of that automatically for us.

Reviewed By: @paulbiss

Differential Revision: D1696229

Signature: t1:1696229:1416537939:b777eeda6f18173c356b950b48eb06cb0e31a7ed