Remove IRREGULAR_INSTRS and PSUEDOINSTRS
Summary: Normalize how everything calls hhbc-translator, so we can generate it.
Some of the blockers to doing this earlier are gone (inlining is no longer in
ir-translator), so most of this is just straight-forward; the one bummer is
some bits of information are still only available in NormalizedInstruction
right now, so we have to make HhbcTranslator have access to it on each
bytecode.  We can gradually weed that out later.

Reviewed By: @swtaarrs

Differential Revision: D1701003