Support Switch/SSwitch vasm instructions in llvm backend
Summary: This diff adds support for enough instructions for the basic switch
statements in the new test case to work in the llvm backend. I added dummy
support for bindjcc2nd since that's used for bounds checking, but it will trap
if the jump is taken (until we can smash jumps).

Reviewed By: @jasone

Differential Revision: D1705102