Fix bundled PCRE build
Summary: Do not pass the "REQUIRED" argument when finding PCRE, since we now have
a bundled PCRE which we can use in the absence of a system library. The
only effect of REQUIRED was to make FIND_PACKAGE_HANDLE_STANDARD_ARGS()
fail with an error.
Closes https://github.com/facebook/hhvm/pull/4342

Reviewed By: @fredemmott

Differential Revision: D1706218

Signature: t1:1706218:1417023737:6d291bcddb6dd21273e12263614e4a59524b3d09