Fix predicted local type invariant failure (test/slow/ext_collator/1810.php)
Summary: It looks like this test is failing in trunk?  This fixes it, and is
definitely a bug fix (because we might only have Gen for the local now even if
we had a LdLoc that knew more earlier, so taking this prediction could lead to
an invariant violation of exactly this sort).  I didn't think through all the
details of why exactly it failed though.

Reviewed By: @markw65

Differential Revision: D1708111

Signature: t1:1708111:1417210961:f0d864b7a9cd844d83f21411e050079150f3255b