Restructure how catch blocks are normally created
Summary: Changes things so that whenever you generate an instruction that can
throw, normally we automatically make a catch block for it (using the state at
that program point).  Adds a variety of assertions try to verify that nothing
has changed the stack before we can throw unintentionally, since we're going to
spill the stack to look like it did before the HHBC opcode started executing by
default.  These assertions are also on for custom catch blocks, which are
needed in irgen-minstr and irgen-builtin, so they have to inform IRBuilder if
they want to modify the stack before a PEI.

This should fix catch blocks to get the correct frame-state as their block in
state in, generally (except for the stacks-states-arent-merged issue).  This
should also make it possible to fix those simplify.cpp issues where we needed
extra catch blocks and didn't have them, but this diff doesn't do that (after
this diff we should figure out whether simplify should be allowed to reuse
catch blocks at all---I am thinking it probably shouldn't be allowed to,
because the block could need a totally different in-state for the new position
we're generating the instruction, and if it made decisions at gen time based on
the previous state it could be wrong).

Reviewed By: @swtaarrs

Differential Revision: D1715225