Fix some bugs when CreateCont's FunctionSuspendHook throws
Summary: When the JIT is on, it leaks the continuation object.  With or without
the JIT on, it was tearing down a frame had already been teleported to the heap
(e.g. the ExtraArgs struct).  (With EnableObjDestructCall on this should
presumably be able to segv by double-freeing---but barring that I'm not sure
how the generator would be reachable yet.)

Unfortunately fixing the leak confuses refcount opts into thinking there /is/ a
leak.  I'm not sure whether we still inline "outer generator" functions---if
not it's not worth running refcount opts in any translations that contain
CreateCont.  Even if we do, we have way fewer generators than we used to, so
let's just give up on the pass when we see this opcode.

Reviewed By: @swtaarrs

Differential Revision: D1709115