Remove catch blocks that pass 'extra spills' to makeCatch
Summary: The makeCatch with extra spills is suspect, because if you spill to a
different depth than your marker, the unwinder won't see the values.  After a
bit of digging around, the current use cases started to make some sense to me,
though: it looks like we're mostly using this trick to tell things to the
refcount validator about what values are consumed on the catch (and in the case
of FPushClsMethod it's actually kinda to lie to it, because the stores it sees
are not actually visible to the unwinder, because they're below the top of the
stack, and we actually do end up with a mismatched stack depth after its
SpillStack).

I don't think that this diff is a fundamentally better way to do this vs doing
this work in catch traces somehow.  And I could also imagine us switching back
again later if we change things so we always re-enter at the maxStackCells
depth (so we don't need the special updateMarkers when we're using prelive
actrecs before advancing bcOff, among other things).  Also, after this change
the unwinding code here is still much more subtle than I'd like (although maybe
in slightly fewer ways---since only the updateMarker/arPreliveOverwriteCells
mechanism is going on now).

But I think that there are a couple things to recommend the change for now,
though: the fact that sometimes the makeCatch 'extra spills' made things
visible to the unwinder and sometimes it didn't seems too surprising---either
it should always work or we shouldn't have it---so this diff removes the whole
mechanism; putting these stores in C++ should save a marginal amount of
generated code size in unwind blocks (meh); and most relevant for what I'm
doing: making these catch blocks 'simple' fits them more easily into the
automatic-generation of catch blocks thing, which we need to make catch block
in-states (except stacks for now) always come from right before the
instructions that actually can throw.

Reviewed By: @markw65

Differential Revision: D1709295