Remove NewPackedArray (and fpushCufArray optimization), neither happen anymore
Summary: This code is dead, since NewPackedArray is now only used for arrays
larger than kPackedCapCodeThreshold (which is huge).  We probably should
consider reimplementing the fpushCufArray optimization to work with the new
InitPackedArray scheme (we may need to add some kind of BlessPackedArray opcode
to bag up the values if we want to do the optimization at parse time still).
That seems kinda like a lockdown thing---I will make a task and reference this
diff as a starting point/reference if folks are ok with removing it for now.

Reviewed By: @markw65

Differential Revision: D1709296

Signature: t1:1709296:1417725906:8315ff5dbeb5ff536575d253e1e626294633e5f1