Only compile ClsCnsD into literals for persistent classes
Summary: If the class is not persistent or a context parent, we're most likely
in non-repo mode (sandboxes), where this could lead to code that always
side-exits if you're editing the file that defines the class.  Mainly removing
this is just to simplify the code though (this code path is difficult to test
and shouldn't buy anything for repo mode, so let's remove it).

Reviewed By: @bertmaher

Differential Revision: D1709411

Signature: t1:1709411:1417642010:81e0c256e305d96559e94bf4c5388397caaa8312