Fix bug relating to m_curWhere in ir-builder (by removing m_curWhere)
Summary: When you pushed a block without providing an iterator, it would set
m_curWhere to the back of a block instead of using folly::none, which was
presumably just a bug.  (It breaks things when appendInstruction needed to make
new blocks.)  But apparently nothing needs this curWhere stuff after D1709457,
so instead of fixing it we can just delete it.

Reviewed By: @swtaarrs

Differential Revision: D1709556

Signature: t1:1709556:1417468647:69af083ae52e3645856286a7526d37ad70471d57