Fix Memcached::setOption for OPT_LIBKETAMA_COMPATIBLE
Summary: The current implementation is missing an if-statement which causes
the option to be unset even when a truthy value is passed in. The comment
directly above this block describes the correct behavior. See the original PHP
implementation for reference: http://goo.gl/7p7WqP. A new constant
Memcached::DISTRIBUTION_CONSISTENT_KETAMA is added for the sake of testing.