Eliminate some stores in pushVMState
Summary: gcc wasn't able to do this optimization automatically, so it was
writing the VMState object to the stack then copying it to the TinyVector. I also
tried adding emplace_back() to TinyVector and using that, but the generated
code wasn't quite as compact as this solution.

This makes VM reentry cheaper and reduces (but doesn't eliminate) the effect of
D1665434.

Reviewed By: @jasone

Differential Revision: D1713106