Smarter path checking in out2expectf
Summary: The old method assumed that file paths would contain /hphp, and
stripped everything from the leading / to the hphp. Some tests end up
with the path to the repo in them, which probably doesn't contain hphp,
or might contain things like hphp-1, hphp-2 etc. In either case,
out2expectf produces expectf files that work locally, but probably not
anywhere else.

This finds the path to the root of the repo, and strips it explicitly,
along with a /hphp if its there, replacing the whole lot with %s.

Reviewed By: @swtaarrs

Differential Revision: D1713663

Signature: t1:1713663:1417626839:d95ef223421ebd9ec6cb0ddd70ef19421f772c62