Tiny amount of easy NormalizedInstruction-removal in irgen-minstr
Summary: Everything easy, so not ni.inputs and ni.imm, and the call to
getPropertyOffset (uses ni.inputs).  It seems like ni.inputs is actually harder
to use than manually running through the mvector to me these days (it has to
build a hashtable for stack offsets and stuff), but removing that use is going
to need a little more restructuring of the code.  Also ni.imm is a tad
inconvenient to remove at the moment because the final op functions don't get
the unpacked immediates...

Reviewed By: @alexmalyshev

Differential Revision: D1715895

Signature: t1:1715895:1417640171:c13ac5ba6e4f1b27fa668789b82f8e841993feb5