Enable a few more assertions for catch trace states and startBlock
Summary: Now we should never be calling startBlock on a non-empty block, and we
can assert that catch traces have in-states that are the same as what they
would be at the instruction that has them as a taken() block.

We can't make this assertion for exit blocks yet (emitSetL fails it, for
example, but only because it's done loads of locals before the instructions
that might exit).

Reviewed By: @alexmalyshev

Differential Revision: D1717822