Use make_shared instead of shared_ptr constructor in getLocation
Summary: With ASAN enabled hphp dies trying to build www, with the error:
```
The process has exhausted 8192MB for size class 48.
```

The culprit turns out to be shared_ptr control blocks, since we have a ton of
shared_ptrs. Using make_shared allocates the object and control block in a
single allocation, so it spreads the size classes around.  This is probably a
good change to make everywhere, but getLocation by itself accounts for like 40%
of the shared_ptrs.

Reviewed By: @edwinsmith

Differential Revision: D1716116