Fix load-elim behavior with type reflows
Summary: Change it so it reflows types after the pass, and inserts AssertTypes
instead of Mov instructions.  Now we can eliminate loads as long as the tracked
type is a subtype of the type of the dst->type() of the instruction we want to
replace, and it shouldn't be able to run into the copyTV issue.

Reviewed By: @jasone

Differential Revision: D1720342

Signature: t1:1720342:1417734422:e0085c8c43aeeb3e8698d9c6e0e538b91c4033b3