Fix issues with how simplifyMod handles its undefined inputs
Summary: Mod has undefined behavior for 0 or -1 in the rhs, but I don't think
this makes it ok for the optimization in simplifyMod to have undefined
behavior.  If we're in unreachable code, we may very legitimately create IR
that would have undefined behavior at runtime, but we shouldn't break while
analyzing or compiling it.  So we can't assert it's not -1 or 0, we have to do
something.  Also remove an old block of code that was working around some
IRBuilder::cond bug that was probably fixed ages ago.

Reviewed By: @ottoni

Differential Revision: D1721311

Signature: t1:1721311:1417816128:7d4152daf4625498edd60d3fe2d6f01766ab6f59