Add ICU DateTimePatternGenerator API to HPHP
Summary: Added a PHP API for the C++ icu::DateTimePatternGenerator. I tried to follow the conventions of the other ICU PHP classes. When there is an error, we throw an exception. The error message/code can be accessed through getErrorMessage/getErrorCode. Like many other Intl classes, the way to create a DatePatternGenerator is through a static method, createInstance or createEmptyInstance.

Reviewed By: @bgertzfield

Differential Revision: D1691412