Remove the SoftClosureReturnTypeHints runtime option (part 1)
Summary: The SoftClosureReturnTypeHints runtime option was added as a temporary
crutch while adding return type enforcement for closures. We don't need it
anymore, so let's get rid of it.

This diff removes all uses of the SoftClosureReturnTypeHints runtime option
inside HHVM, but doesn't completely remove the option because there are
still some configs that reference this option. Later, we'll completely
the option once enough time has passed to update all the relevant configs.

Reviewed By: @elgenie

Differential Revision: D1724592