Add VectorFormat to unserialize a vector as map or as vector
Summary: The current code serialize vector as map. But in theory, we should serialize the vector as a vector.

This diff adds a new FB_SERIALIZE_VECTOR used for vectors.

The serialize logic now can serialize the vector as
FB_SERIALIZE_VECTOR
list of Values
FB_SERIALIZE_STOP

The unserialize logic will understand the FB_SERIALIZE_VECTOR correctly, and use createVector and vectorAppend to create the map (php doesn't have vector type so it still uses map).

To make roll out eaiser, this diff only has the unserialize part. The serialize part will be checked in after unserialize it everywhere.

Reviewed By: @andriigrynenko

Differential Revision: D1714262