Improve thread hooks
Summary: Also fix a couple of leaks on thread exit.

I don't think we're being bitten by the leaks right now, because
the only threads that are created/destroyed a lot are created using
std::thread, which bind directly to pthread_create etc, bypassing
our wrappers.

But I didn't like the way we had to sleep/retry to get hold of the
thread info from the map. So this passes the PthreadInfo as the
argument to the start function.

Reviewed By: @jasone

Differential Revision: D1731592

Signature: t1:1731592:1418245780:1ef8a7414641474036d0c76d3fc875693ca55ec1