Make timeout flag sticky
Summary: Occasionally we see long-lived requests that fail to timeout.  What
probably happens is that the TimeoutException is generated, the timeout
surprise flag is unset, and then the exception is caught by some block that
simply drops it.  While we could keep auditing catch blocks, it seems more
reliable to simply make timeouts sticky, so that we'll re-raise the exception
at the next opportunity.

Reviewed By: @jdelong

Differential Revision: D1711487