fixed decoding utf8 strings so they correctly return '?' on invalid sequences
Summary: fixed utf8 decode function so it will it will correctly handle invalid byte sequences. For example, the the old decoder allowed the non shortest form of a character to be decoded, which allows for a number of malicous attacks. Fixes #4442428, which is CVE-2010-3870 (Invalid UTF-8 passed to utf8_decode results in unexpected output)

Reviewed By: @ptarjan

Differential Revision: D1488851