Implement smashable jump service requests in llvm backend
Summary: Fairly straightforward, except for the fact that the jumps aren't
actually smashable yet due to a bug in llvm's assembler (or how we're using
it). The bindjcc1st/2nd vasm instructions are implemented with a local jcc
going to REQ_BIND_JMP stubs because we can't emit smashable jccs yet.

I moved the global function name table into TCMemoryManager to keep the size
from growing indefinitely with each translation, since we need to generate
names for certain stubs that are unique for each translation.

Reviewed By: @maksfb

Differential Revision: D1717035