Make ast destruction less recursive
Summary: ast nodes can have a pointer to a canonical node which
allows a certain amount of cse and constant propagation. These could
chain through the elements of an array, with the result that when the
array is destroyed, the last element keeps all the other elements alive,
and when its finally destroyed, we recursively walk the canonPtrs
destroying the elements.

This just clears all the canonPtrs in the destructor of ExpressionList,
which should prevent most pathalogical cases of this.

Reviewed By: @jdelong

Differential Revision: D1731406