Explicitly prepareForNextHHBC before doing singleton inlining
Summary: It is getting the correct bytecode offsets right now, but it's sorta
indirectly doing so (it's because of the prepareForNextHHBC call that we do
before any guards, to get them associated with that offset).  If one of the
guards is a stack guard (which I think happens only if I turn off
constrainStack and auto-constrain stack guards), then we might redefine the
stack in between prepareForNextHHBC and a PEI in the inlined singleton code,
which is an assertion failure about possibly throwing an incorrect stack state
now.  Fix this by explicitly setting the next hhbc stuff up before we try to
inline.

Reviewed By: @markw65

Differential Revision: D1731989