Don't allow load-elim to reuse values when they have Type::Bottom
Summary: When we analyze a CheckTypeMem, we intersect its typeParam with the
knownType we're tracking.  If the CheckTypeMem is always going to fail this can
leave the knownType as Bottom.  We could try to eliminate the fallthrough in
this case, but for now let's just make sure we don't break IR rules by
generating an AssertType<Bottom> later.

Reviewed By: @alexmalyshev

Differential Revision: D1732250

Signature: t1:1732250:1418252941:e42d488ff67adf9d8fe011e2e0c2ffe0daddb979