Delete interpreter type-profiling code
Summary: This isn't doing anything now that we PGO everything, and it added a
fair bit of complexity to some jit functions. This doesn't touch the "runtime
type profiler" which is something unrelated to type prediction in the jit, and
the non-type warmup profiling for Func hotness and InstanceBits is still
around.

Reviewed By: @alexmalyshev, @jdelong

Differential Revision: D1732733

Signature: t1:1732733:1418259277:acffb558bef3ccc0d1b1ca24bee8918b5e5a9197