Track stack slot states as memory, instead of chasing StkPtrs
Summary: Replaces getStackValue with effectful state-tracking, the way we do
for locals.  After this diff, we still redefine StkPtrs in a few situations:
after calls, because the cross-region ABI still requires this, and at BCCF
merge points (necessary because we might have a call only on one side).  A few
instructions involved in return sequences also do this.

The state for handling this at gen time still looks a lot like it did with
spillStack.  We can't rely on a later pass to eliminate loads and stores to the
stack yet for two reasons: 1) load-elim and store-elim don't know about the
stack yet (we need to teach memory-effects and AliasClass about it), and 2)
refcount opts will break if we don't do the gen-time spillStack-style load
elimination, because it will think the stack is consuming references too much.

This diff also breaks SpillFrame-elimination in optimizeActRecs, because
SpillFrame becomes essential.  I'll plan on getting that to work in store-elim
soon instead.

Reviewed By: @swtaarrs

Differential Revision: D1752689