Fix JIT crash related to stack size limits.
Summary: Fix the JIT to distinguish between the native stack and the VM stack,
even when the VM stack ends up allocated in the virtual memory region
that was originally available to the native stack.  This situation
arises when the stack size limit (ulimit -s) is set very high, e.g.
unlimited, in which case shared libraries are loaded much lower in
memory in order to give the native stack lots of room to grow.

Reviewed By: @swtaarrs, @markw65

Differential Revision: D1736940