Disable fewer warnings for typing/
Summary: This is still a work-in-progress: I was originally going to
disable almost all the warnings, but there seems to be a bunch of unused
code of possible value (e.g. CompareTypes) that I'm not 100% sure that I
should delete. Will ask@pikatchu on Monday. But the work thus far is
worth committing.

Reviewed By: @jwatzman

Differential Revision: D1739224

Signature: t1:1739224:1418678840:66580352d133a9a27910a24569b350d3c4e0ea47