Reduce dependencies on ParserHeap.
Summary: The goal is to allow us to discard parsing data after our types
are declared, so we can reduce the size of the saved state. To this end,
things that depend on the ParserHeap should be switched to use
files_info wherever possible.

Reviewed By: @jwatzman

Differential Revision: D1739737

Signature: t1:1739737:1418683152:9388580dbef61b2aac5cc601cfc4db67deb0760b