Don't sink IncRefs past SyncABIRegs
Summary: SyncABIRegs is supposed to ensure cross-tracelet ABI registers are in
the proper state for leaving the trace, so we shouldn't have anything between
it and the exit following it. We've never verified or depended on this, but the
llvm backend does check some aspects of it now, and this was causing a failure
in one repo mode test.

Reviewed By: @ottoni

Differential Revision: D1739443

Signature: t1:1739443:1418667644:9ea10fe09915ae0fb55215ff22d0b656a14db7e2