Fix some test*im operand sizes
Summary: We were using cmpqim and testlim on the same memory address, first to
check a pointer for null and then to inspect something in its bottom half. This
is only ok on little-endian machines, and the LLVM optimizer will do funky
things to endianness-dependent code. I shuffled things around so we always emit
a properly-sized test instruction in code-gen-x64.cpp, and vasm-x64.cpp will
emit a smaller instruction when appropriate. I know the instruction-shrinking
logic in vasm-x64.cpp could be smarter than it is; I decided to keep it exactly
as smart as it was before for simplicity.

Reviewed By: @jasone

Differential Revision: D1737884

Signature: t1:1737884:1418431658:638373b0d36ab6e731513d355c9306c93c121154